requirements.txt
torch>=2.0.0
torchvision
torchaudio
numpy
pandas
scipy
matplotlib
seaborn
tqdm
einops
Pillow
